+++ src/z3c/autoinclude/utils.py
@@ -39,7 +39,7 @@ class ZCMLInfo(dict):
         for zcml_group in zcml_to_look_for:
             self[zcml_group] = []
 
-    
+
 def subpackageDottedNames(package_path, ns_dottedname=None):
     # we do not look for subpackages in zipped eggs
     if not isUnzippedEgg(package_path):
@@ -69,8 +69,9 @@ def distributionForPackage(package):
 
 def distributionForDottedName(package_dottedname):
     """
-    This function is ugly and probably slow.
-    It needs to be heavily commented, it needs narrative doctests, and it needs some broad explanation.
+    This function is ugly and probably slow. It needs to be heavily
+    commented, it needs narrative doctests, and it needs some broad
+    explanation. It fails in some namespace cases (see utils.txt).
     Then it needs to be profiled.
     """
     valid_dists_for_package = []
@@ -89,9 +90,9 @@ def distributionForDottedName(package_dottedname):
 
     if len(valid_dists_for_package) == 0:
         raise LookupError("No distributions found for package `%s`; are you sure it is importable?" % package_dottedname)
-    
+
     if len(valid_dists_for_package) > 1:
-        non_namespaced_dists = filter(lambda x: len(x[1]) is 0, valid_dists_for_package)
+        non_namespaced_dists = list(filter(lambda x: len(x[1]) is 0, valid_dists_for_package))
         if len(non_namespaced_dists) == 0:
             # if we only have namespace packages at this point, 'foo.bar' and 'foo.baz', while looking for 'foo',
             # we can just select the first because the choice has no effect
@@ -126,7 +127,7 @@ def namespaceDottedNames(dist):
     except KeyError:
         ns_dottednames = []
     return ns_dottednames
-    
+
 def isUnzippedEgg(path):
     """
     Check whether a filesystem path points to an unzipped egg; z3c.autoinclude